Rewrote Makefile to Rakefile
Looks better, works better, easier to modify.

Moved scp file transfer command to external file src/publish.sh to avoid showing the path of website on the server. It can be called from Rakefile, if needed (commented out to avoid problems for persons who don't have it yet). Also added the filename to .gitignore so that it wouldn't get added to git repository.